Cleans up the way data is saved
Note that a whole new folder, ui_tools_private, has been added. This is a copy of the private folder in ui_tools, because I wanted to make one tiny change to how msgbox works. The extra file in here that we don’t need should probably be removed, and the ones we DO use should be renamed